requirements.txt
numpy
pandas
snowflake-connector-python
streamlit
